Oil Rig Sim Complete! I win!
Oil Rig simulation including water chem yields a turtle line that is relatively flat even when Sea Pearls (the basis of the coral food chain) are removed. This leads to pH doing wonky things and the coral dying out. The turtles just keep on keeping on as do the cyanobacteria which actually thrive in the absence of the sea pearls.